First commit. Updated the source data and did some trivial fixes.

+++ include/LinearRegression.h
@@ -4,7 +4,7 @@
 #include <vector>
 
 struct StockData {
-    double date;
+    int date;
     double price;   
 };
